Feature/validation des donnees cote back

+++ back/src/main/java/com/crystalfinance/api/controller/CoordinatesController.java
@@ -1,11 +1,14 @@
 package com.crystalfinance.api.controller;
 
+import com.crystalfinance.api.exception.CoordinatesNotFoundException;
 import com.crystalfinance.api.model.Coordinates;
 import com.crystalfinance.api.service.CoordinatesService;
 import com.crystalfinance.api.service.DistanceCalculateService;
 import com.crystalfinance.api.utils.CoupleOfId;
+import jakarta.validation.Valid;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -28,28 +31,36 @@ public class CoordinatesController {
     @GetMapping("/coordinates")
     public ResponseEntity<List<Coordinates>> getCoordinatesList() {
         List<Coordinates> coordinatesList = coordinatesService.getCoordinatesList();
-        log.info("providing coordinates list");
         return ResponseEntity.ok(coordinatesList);
     }
 
     @PostMapping("/coordinates")
-    public ResponseEntity<Coordinates> addCoordinates(@RequestBody Coordinates coordinates) {
-        Coordinates addedCoordinates = coordinatesService.addCoordinates(coordinates);
-        log.info("adding new coordinates");
-        return ResponseEntity.ok(addedCoordinates);
+    public ResponseEntity<Object> addCoordinates(@Valid @RequestBody Coordinates coordinates, BindingResult result) {
+        if (result.hasErrors()) {
+            throw new CoordinatesNotFoundException("received coordinates are not valid");
+        } else {
+            Coordinates addedCoordinates = coordinatesService.addCoordinates(coordinates);
+            return ResponseEntity.ok(addedCoordinates);
+        }
     }
 
     @DeleteMapping("/coordinates/{id}")
     public ResponseEntity deleteCoordinates(@PathVariable("id") String id) {
-        Coordinates deletedCoordinates = coordinatesService.deleteCoordinates(id);
-        log.info("deleting coordinates");
-        return ResponseEntity.ok(deletedCoordinates);
+        if (id.isBlank()) {
+            throw new CoordinatesNotFoundException("received id is not valid");
+        } else {
+            Coordinates deletedCoordinates = coordinatesService.deleteCoordinates(id);
+            return ResponseEntity.ok(deletedCoordinates);
+        }
     }
 
     @PostMapping("/couple-of-id")
-    public ResponseEntity<Long> getCoupleOfId(@RequestBody CoupleOfId coupleOfId) {
-        distance = distanceCalculateService.calculateDistance(coupleOfId.getIdOne(), coupleOfId.getIdTwo());
-        log.info("providing distance");
-        return ResponseEntity.ok(distance);
+    public ResponseEntity<Long> getCoupleOfId(@Valid @RequestBody CoupleOfId coupleOfId, BindingResult result) {
+        if (result.hasErrors()) {
+            throw new CoordinatesNotFoundException("received ids are not valid");
+        } else {
+            distance = distanceCalculateService.calculateDistance(coupleOfId.getIdOne(), coupleOfId.getIdTwo());
+            return ResponseEntity.ok(distance);
+        }
     }
 }
